+++ test/data/test2.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+read value
+echo "value: $value"