+++ test/data/test3.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+# used to validate _password
+echo "\$QS_password: $QS_password"
+echo "\$QS__password: $QS__password"
+echo "\$QS___password: $QS___password"